+++ CHANGELOG.md
@@ -3,5 +3,5 @@
 
 - added built-in `quit` as an alias for `exit`, 16 Sept 2023, George Dance
 - added simple `Makefile` to automate building, 16 Sept 2023, George Dance
-
+- added built-in `prompt` to change the prompt, 16 Sept 2023, George Dance